refactor: :recycle: seperate themes from common
