+++ includes/header.ejs
@@ -1,4 +1,4 @@
-<script type="module" src="/resources/js/common.js?<%- fileHashes.js['common.js'] %>"></script>
+<script type="module" src="/resources/js/themes.js?<%- fileHashes.js['themes.js'] %>"></script>
 <header>
   <div data-tippy-content="Supporting LGBTQ+ rights all around the year! ❤️<br><br>Click for a surprise. :)" class="pride-flag logo"></div>
   <a href="/" id="site_name">SkyCrypt</a>
