+++ includes/resources.ejs
@@ -50,5 +50,7 @@
 
   <script defer type="module" src="/resources/js/common-defer.js?<%- fileHashes.js['common-defer.js'] %>"></script>
 
+  <script src="/resources/js/common.js?<%- fileHashes.js['common.js'] %>"></script>
+
   <script> const page = "<%- page %>"; </script>
   <script> const extra = JSON.parse(`<%- JSON.stringify(extra).replace(/\\/g, '\\\\') %>`); </script>
